DABOM-395 feat: 빌드 에러 수정
