+++ packages/shared/src/components/Badge.tsx
@@ -1,17 +1,6 @@
 import { ReactNode } from 'react';
 
-export type BadgeColor =
-  | 'primary'
-  | 'white'
-  | 'gray'
-  | 'outline'
-  | 'primary_light'
-  | 'negative'
-  | 'gray_light'
-  | 'emergency'
-  | 'approved'
-  | 'rejected';
-export type BadgeSize = 'sm' | 'lg' | 'md_fixed';
+import { BadgeColor, BadgeSize } from '../types/badge';
 
 interface BadgeProps {
   children: ReactNode;
