+++ packages/shared/src/types/badge.ts
@@ -1,0 +1,13 @@
+export type BadgeColor =
+  | 'primary'
+  | 'white'
+  | 'gray'
+  | 'outline'
+  | 'primary_light'
+  | 'negative'
+  | 'gray_light'
+  | 'emergency'
+  | 'approved'
+  | 'rejected';
+
+export type BadgeSize = 'sm' | 'lg' | 'md_fixed';
